Move ApplicationIdModule to common
