+++ paymentsheet/src/main/java/com/stripe/android/link/injection/NativeLinkDIConsts.kt
@@ -1,4 +1,3 @@
 package com.stripe.android.link.injection
 
-internal const val APPLICATION_ID = "application_id"
 internal const val START_WITH_VERIFICATION_DIALOG = "start_with_verification_dialog"
